Support rendering resource changes in progress display
This adds a "ShowResourceChanges" option to the progress display, which when set
adds a section showing resource diffs. The updated test data shows examples.
This combines the tree view with more detailed information about the changes
made, albeit more compactly. This change is intended to be consumed by Pulumi
Copilot, and so is not enabled for the CLI at this time.