Respect .node-version and .nvmrc files in `pulumi install` (#17060)
When using the `--use-language-version-tools` flag with `pulumi
install`, and `fnm` is installed, we look for `.nvmrc` and
`.node-version` files and install the requested version.

This is meant to be used within containers where we don’t run within a
shell. When running locally, fnm, nvm and similar tools manage the
nodejs version via their shell integration and ensure the correct
version is set.

pulumi-docker-containers makes use of this in
https://github.com/pulumi/pulumi-docker-containers/pull/253