Refactor and fix bug in progress_test  (#17079)
Discovered while refactoring, the `raw` argument supplied to a test
utility in the `progress_test` tests was not used.

Fixes the tests in two commits, both of which pass tests. The first
implements the intended refactoring without changing the test behavior.
The second updates the test to use the `raw` parameter and regenerates
test output.

The refactoring is applied in this pull request as a "friendly
amendment" to improve the code before making larger changes to add more
testing options.