Add workspace.Context for mocking workspace functions (#17089)
Prompted by trying to write some more unit tests for `state delete` and
hitting the sadness again of how little of the system is interface'd and
thus can be mocked out in tests.

This is a start on adding an interface for "workspace" related
functions. That's things like loading Pulumi.yaml files, listing plugins
etc.

This PR simply adds the minimal interface for a `ReadProject` function
and replaces the current function for that in pkg/cmd/util with an
implementation in pkg/workspace. All calls of the old `readProject` are
replaced with `workspace.Instance.ReadProject`. The intention is to lift
this parameter up so the only place `workspace.Instance` gets referred
to is in the top level command handlers, everything else should use it
via DI so a mock version can be injected in tests.

We had one test that was mocking `readProject` out, and I've replaced
that to just use `workspace.Context` instead.

These changes are fairly simple, but they touch a lot of lines of code
so I'm breaking this work into lots of separate PRs like this. After
this merges I'll do a pass to lift the `workspace.Instance` access to
just top level, then look at other workspace functions needed like
getting the current backend (which requires accessing user config).