Add LoginManager to backend package
This abstracts the code from `currentBackend` and
`nonInteractiveCurrentBackend` into a new interface `LoginManager`. This
happens to very closely match the `LoginManger` interface defined in the
httpstate backend, but this is intended to be used for diy state as well.

This lays the path to getting rid of backendInstance, which should allow more
tests to run with `t.Parallel()` (plus its just nice to have less global
varaibles).