Add loader_target to RunRequest (#17077)
This adds a schema loader target to the language run request, similar to
the generate project/package requests. This allows a language runtime to
connect back to the engine to load schema information in _exactly_ the
same way as the engine.

We'll want to switch YAML over to using this loader so that it will work
with paramaterized packages and conformance tests.