Make a PULUMI_BACKEND_URL an env.Var instance (#17091)
Looking into writing some mocked tests for `state destroy` this env var
is touched on the login path and is currently not in our "env" system so
tricky to mock out. This is a small PR to just lift it into the env
system without worrying about plumbing the env instance all the way down
to where it's used (yet).

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>