Copy Parameterization to the codegen package (#17098)
We want to be able to pass parameterisation details to GetSchema for
loading parameterised schemas over grpc. This is exactly the same
details we pass to RegisterPackage for parameterised details. We
duplicate the type as we can't refer to it because of import cycles, and
we can't move it because things already use the existing type.