Add GetStoredCredentials to pkg workspace (#17106)
Add GetStoredCredentials to the package workspace interface. We can't
remove the old copy of the code in sdk/go/common, but everywhere in pkg
has been updated to use the interface call instead.

This also allowed the writing of a few more test cases for the
`ValueOrDefaultURL` method in the httpstate package.