Include all engine events in the event log (#17101)
In https://github.com/pulumi/pulumi/pull/14607 we introduced filtering
of engine events before they are displayed. However in the process we
also started filtering them from the event log. The event log is used
for two things, the automation API and debugging. The automation API
already ignores events it doesn't know about (and it should because a
new version of the CLI can always introduce new events that aren't
filtered), and for debugging it is very useful to have all events in the
event log.

In fact we are going to introduce a new debugging event, which the
automation API will need (plugins for IDEs are going to be written using
the automation API to intercept the engine events).

Move the filtering after the writing to the event log, so we always have
all events there, but still filter events that are not relevant to the
service etc.