Support loading parameterized schemas in the schema loader (#17108)
This adds new `LoadPackageV2` and `LoadPackageReferenceV2` methods to
the schema loaders. These take more information than just the "name" and
"version" allowing the loading of paramaterized schemas.

Also plumbed the download url though while at it, so if a request comes
in for a plugin not yet fetched the schema loader should be able to
download it.

To help with testing I've added MockHost and MockProvider structs to the
plugin package, and used them to test that the plugin loader can now
load a parameterised package.