Propagate deleted parents of untargeted resources (#17117)
When using `--target` to target specific resources during an update, we
use the list of targets to decide which steps to generate given a set of
resource registrations. Specifically, *prior to #16247*:

* If a registration event named a resource that was targeted, we would
process it as usual.
* If a registration event named a resource that was not targeted, we
would emit a SameStep for it.

In #16247, this behaviour was changed so that we'd consider the entire
dependency graph of an untargeted resource, in order to avoid missing
copying resources that had been removed from the program but not
targeted (and thus whose old states would be copied over later on,
potentially *after* the states of resources depending on them). As part
of this work, parent-child relationships were not handled. This did not
manifest in errors in the cases tested, but in the presence of aliases
and other edge cases, parent-child relationships *can* result in
snapshot integrity issues and thus must be handled too. This commit
implements those changes.

Fixes #17081 (I believe)