Switch back to `os/user` to get the current user
This change switches back to using the stdlib's `os/user` to get the current user. We originally used `os/user` but moved to a drop-in replacement (`github.com/tweekmonster/luser`) in #5065 which worked better without `cgo` at the time. However, `github.com/tweekmonster/luser` no longer appears to be maintained and the stdlib's `os/user` has a pure Go implementation (which doesn't require `cgo`) which has improved since then, so we should be able to switch back to it now.