ResourceRefsGetResourceGo: also wait for the childs message to be registered (#17116)
In https://github.com/pulumi/pulumi/pull/16897 we made sure we wait for
the child to be registered in the container before we continue to avoid
racyness. We however also need to do the same for the message registered
on the child, otherwise that can still race. Do so here.

Fixes https://github.com/pulumi/pulumi/issues/17095