wire up command line interface for running programs under a debugger (#17087)
After adding the engine bits for debugging in a previous PR
(https://github.com/pulumi/pulumi/pull/17072), implement the option in
the CLI that triggers it. This by itself will have no effect still until
the language runtimes have support for it. Therefore we keep it disabled
until we've implemented this in the respective language runtimes.

The language runtime implementations will follow in subsequent PRs. I'm
a bit unsure about the naming of the flag. Maybe `--run-in-debugger`
would be a better name? Would be great to get some opinions on that (or
any other suggestions fi someone has an idea).