plugins: don't panic when the given path is not clean (#17141)
When the user specifies a provider with a specific path in the
`Pulumi.yaml`, we later in the program assert that the path that was
passed in also matches with the path where we found the plugin. We do
this by using a string comparison, however that doesn't work if the path
the user passes is not clean, e.g. has a trailing slash, or has a double
slash, or some such.

Fix this by using `filepath.Clean` on the user supplied path, which
these things up.

I was also briefly wondering if this works properly if the user passes
in a path that is a symlink (it does), and wrote a test for that,
checking that behaviour.

Fixes https://github.com/pulumi/pulumi/issues/17130