debugging: more consistency for naming (#17152)
We're a bit inconsistent with naming for debugging, calling it enable
debugging in some places, attach debugger in others, and start debugger
yet in others.

Unify this, so we call everything that goes from the engine towards the
language host attach debugger, while the message we get back from the
language host is called start debugging since we're telling the engine
that we want to be starting the debugging session.

Thanks @julienp for the suggestion!