run acceptance tests for windows on PR rather than on merge (#17114)
Acceptance tests on Windows sometimes fail for mysterious reasons. It's
unclear to me whether this is because the tests sometimes genuinely take
too many resources, and break the windows host, or whether it's
instability of the windows CI runners in general (it could also be
both).

This is very annoying during merges, because for any failed test we need
to re-do the merge, which means re-running all tests, which is both
expensive and slow.

Instead, run the acceptance tests on windows during PRs always, but
don't run them during merge.

I think this is okay to do because if there is a mismerge, I don't
expect exclusively windows tests to fail, but I expect that to also
affect Linux and MacOS runners. It seems very unlikely for a Windows
specific problem to pop up only when multiple PRs are merged together.

It is also convenient to run the windows tests during PRs in the first
place, because those are tests developers usually don't run on their own
machine, so it's not super unlikely they fail. We used to not run these
tests on PRs because it's kind of expensive to run them, but we have
more headroom now, so running them during PRs should no longer be a big
issue.