Improve snapshot integrity panics depending on operation (#17144)
Snapshot integrity errors generally happen for one of two reasons:

* We just *wrote* a snapshot with an issue.
* We just attempted to *read* a snapshot with an issue.

As much as possible, we want to capture errors and issues at the point
of *writing* the snapshot, since that will generally be where the bug
that caused the integrity issue lies. This commit updates snapshot
integrity errors to carry awareness of whether they happened during a
read or write, and modifies the panic message to nudge users to hunt
down the failing write operation if the panic they are seeing relates to
a read.


![image](https://github.com/user-attachments/assets/6d91b5ea-95f5-470b-992f-1abfa116c838)

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>
Co-authored-by: Eric Rudder <ericrudder@users.noreply.github.com>