add debugging support to Python language host
Add debugging support for the Python language. The scaffolding for
this was put in place previously in #17072 and then #17087.

To do this we start the program under debugpy, and then wait for the
debugger to get ready before informing the engine that it needs to
inform the user that a debugger should be attached.