Add a `test_lifecycle` target to `Makefile` (#17173)
Lifecycle tests (located in `pkg/engine/lifecycletest`) allow us to
exercise the various engine code paths controlling the resource
lifecycle -- step generation, step execution, and so on. It's often
useful when working in the engine (e.g. hunting down a snapshot
integrity issue) to just run the lifecycle tests. This commit adds a
target to the `Makefile` to do just that.