add the attach-debugger flag to automation API (#17175)
The VSCode extension will make use of the automation API, so we'll need
to be able to pass the flag through there. While the VSCode extension
will only use the TypeScript API, let's add the flag to every language
in case someone finds a use for it.

Fixes https://github.com/pulumi/pulumi/issues/17174