Correctly case Python package names in `package add` link instructions
When we print out instructions for linking a local Python SDK in
response to running `package add`, we incorrectly use the Pulumi package
name, which may not be a valid Python package name. This commit fixes
this to use the specified Python package name override if provided, or
the correctly-cased Pulumi package name if not.

Fixes #17149