Export a helper for rendering progress to a "simple" terminal (#17147)
The helper method exported here uses an internal terminal,
"SimpleTerminal", which ignores terminal control sequences and writes
its output directly to a buffer. The wrapper retains a similar API as
`ShowProgressEvents` to enable streaming events, though performance may
be suboptimal as a subset of events must be buffered as rows to render
the tree.

For cleaner output in this mode, the `CarriageReturn` operation is added
to the internal terminal interface to conditionally render it.
Ordinarily this is used along with cursor movement to render the tree in
place, and is not needed for this mode. This removes spurious `\r`
(rendered as "^M") characters from the diff for the plain rendering
tests.