Recover from zero-byte `credentials.json` files
Pulumi stores credentials in a `credentials.json`. If a non-atomic
operation that writes this file is interrupted, we may end up with a
zero-byte `credentials.json` file. Presently, this will create a
situation where the user has to manually remove the file before e.g.
`pulumi login` will work and retrieve new credentials again (see #11017
for an example). This commit changes this behaviour, spotting empty
credentials files and returning an empty set of credentials instead of
throwing an error.

Fixes #11017