add debugging support to Go language host (#17102)
Add debugging support for the Go language. The scaffolding for this was
put in place previously in https://github.com/pulumi/pulumi/pull/17072
and then https://github.com/pulumi/pulumi/pull/17087.

To start the debugging, we're starting the program under dlv, and send
the debugging information to the engine, which will emit an engine event
and print information to the user to allow attaching the
debugger.   For Golang the only thing we're printing for the user is
the port number, which is enough to reconstruct the information to
attach a debugger using the DAP protocol.

E.g. for emacs the following setup is needed to be able to use this in
`dap-mode`:

```
(dap-register-debug-template "Go: Attach to Delve"
  (list :type "go"
        :request "attach"
        :name "Attach to Delve"
        :mode "remote"
        :host "127.0.0.1"
        :port <port>))
```

Fixes #17170

---------

Co-authored-by: Eron Wright <eron@pulumi.com>