Hide unnecessary rows in non-interactive mode
When using a `--non-interactive` renderer, it seems that we don't
respect the presence or absence of options like `--show-sames`, instead
displaying all rows. While we do correctly compute whether or not a row
should be hidden, we don't check that computed boolean later on unless
we are performing an interactive render. This commit fixes that.

Fixes #12162