add debugging support to Nodejs language host (#17156)
Add debugging support for the Go language. The scaffolding for this was
put in place previously in #17072 and then #17087.

To start the debugging, we're passing the `--inspect-brk` option to
node, which enables debug mode, and pauses the program until the
debugger is attached.

Again writing tests for this seems a bit tricky, since we need to attach
to a process (unless someone knows the trick on how to do that?)

Fixes #17171

Co-authored-by: Eron Wright <eron@pulumi.com>