state move: show full set of delete commands when writing source stack fails
Before actually writing the source stack, we have no way of knowing we
can write it.  The same is true for writing the destination stack.
This means if the one we write later fails to write, we get an
incompletely moved state.

We've decided to move the destination stack first, and then the source
stack, so resources are always tracked.  The error message also
mentions cleaning up the source stack, however it's not completely
obvious how to do that.  Since we know the resources that were
supposed to be moved (and successfully been written to the destination
stack at this point), we can give the user better instructions on how
to clean up the resources.

Do that by giving them a full set of `pulumi state delete` commands
they can run to clean up the source stack.