auto: support refresh consistently in up/preview/destroy
Currently refresh support is inconsistent in automation API.  Go
supports it for up, preview and destroy, NodeJS only supports it for
up and preview, and Python doesn't support it at all.  Support it for
all three in up, preview and destroy.

The important bit is supporting it for NodeJS (see https://github.com/pulumi/actions/pull/1278#pullrequestreview-2290185113),
but it's best to support it consistently when we can.