Add LoginManager to backend package (#17199)
This abstracts the code from `currentBackend` and
`nonInteractiveCurrentBackend` into a new interface `LoginManager`. This
happens to very closely match the `LoginManger` interface defined in the
httpstate backend, but this is intended to be used for diy state as
well.

There were a few places using `loginToCloud` to specifically login to
the cloud backend. I've replaced these with just
`DefaultLoginManager.Login` because either they are specifically passing
in a cloud URL, or possibly an empty URL which will in Login select the
cloud backend.

This lays the path to getting rid of backendInstance, which should allow
more tests to run with `t.Parallel()`, and be a more obvious mock point
(plus its just nice to have less global variables).

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>