state move: revert dest changes if source can't be saved (#17208)
As explained in https://github.com/pulumi/pulumi/pull/17205, we need to
save one stack after another when doing a state move. The mentioned PR
introduces a better error output, so the user can get themselves out of
the broken situation.

However we can do a bit better. If we fail to save the source stack for
any reason, we can still try to revert the changes from the destination
stack, bringing us back to the original situation, without the user
having to intervene manually, which is usually preferable for users. If
we fail to do *that*, we still display the list of commands the user can
use to get themselves out of the situation.

Unfortunately this makes that latter output very hard to test, since
we'd have to first succeed in saving the destination stack and on the
next attempt fail, and our test infrastructure is not quite set up for
that.

This is based on https://github.com/pulumi/pulumi/pull/17205, and should
be another improvement to the user experience in that case.

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>