Refer to MLCs as component providers (#17238)
Component providers are those which implement the `Construct` endpoint
to supply remote components. Internally, we often refer to these as
"multi-language components", or MLCs, since historically this was the
point of introducing them -- to enable a program in any language to
consume a component written in any language. That said, "component
provider" is arguably a better term. Firstly, such providers may also
provide other non-component resources, and not "just" be "MLCs".
Secondly, "component providers" is the term we use when interacting with
customers, since it more effectively communicates the use case in
question. This commit reflects this preference by documenting component
providers, and not MLCs, in our developer documentation, fleshing out a
few things (such as local vs. remote) while we're at it.